requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.2.1
    # via jupyter-server
argon2-cffi==20.1.0
    # via
    #   jupyter-server
    #   notebook
async-generator==1.10
    # via nbclient
attrs==21.2.0
    # via jsonschema
babel==2.9.1
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.9.3
    # via -r requirements.in
bleach==3.3.1
    # via nbconvert
cachetools==4.2.2
    # via -r requirements.in
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via argon2-cffi
charset-normalizer==2.0.3
    # via requests
cramjam==2.3.2
    # via fastparquet
cycler==0.10.0
    # via matplotlib
debugpy==1.3.0
    # via ipykernel
decorator==4.4.2
    # via
    #   ipython
    #   networkx
defusedxml==0.7.1
    # via nbconvert
easydict==1.9
    # via -r requirements.in
entrypoints==0.3
    # via nbconvert
fastparquet==0.7.0
    # via -r requirements.in
fsspec==2021.7.0
    # via fastparquet
grpcio==1.38.1
    # via pydgraph
hiredis==2.0.0
    # via redisbloom
idna==3.2
    # via
    #   anyio
    #   requests
ipykernel==6.0.2
    # via notebook
ipython==7.25.0
    # via
    #   ipykernel
    #   jupyterlab
ipython-genutils==0.2.0
    # via
    #   jupyter-server
    #   nbformat
    #   notebook
    #   traitlets
jedi==0.18.0
    # via ipython
jinja2==3.0.1
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   notebook
json5==0.9.6
    # via jupyterlab-server
jsonschema==3.2.0
    # via
    #   jupyterlab-server
    #   nbformat
jupyter-client==6.1.12
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
    #   notebook
jupyter-core==4.7.1
    # via
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbconvert
    #   nbformat
    #   notebook
jupyter-server==1.9.0
    # via
    #   jupyterlab
    #   jupyterlab-server
    #   nbclassic
jupyterlab==3.0.16
    # via -r requirements.in
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-server==2.6.1
    # via jupyterlab
kiwisolver==1.3.1
    # via matplotlib
markupsafe==2.0.1
    # via jinja2
matplotlib==3.4.2
    # via -r requirements.in
matplotlib-inline==0.1.2
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
nbclassic==0.3.1
    # via jupyterlab
nbclient==0.5.3
    # via nbconvert
nbconvert==6.1.0
    # via
    #   jupyter-server
    #   notebook
nbformat==5.1.3
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.1
    # via nbclient
networkx==2.5.1
    # via -r requirements.in
notebook==6.4.0
    # via
    #   -r requirements.in
    #   nbclassic
numpy==1.21.1
    # via
    #   -r requirements.in
    #   fastparquet
    #   matplotlib
    #   pandas
    #   pyarrow
packaging==21.0
    # via
    #   bleach
    #   jupyterlab
    #   jupyterlab-server
pandas==1.3.0
    # via
    #   -r requirements.in
    #   fastparquet
pandocfilters==1.4.3
    # via nbconvert
parse==1.19.0
    # via -r requirements.in
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.1
    # via matplotlib
prometheus-client==0.11.0
    # via
    #   jupyter-server
    #   notebook
prompt-toolkit==3.0.19
    # via ipython
protobuf==3.17.3
    # via pydgraph
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pyarrow==4.0.1
    # via -r requirements.in
pyasn==1.6.1
    # via -r requirements.in
pycparser==2.20
    # via cffi
pydgraph==21.3.0
    # via -r requirements.in
pygments==2.9.0
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pyhumps==3.0.2
    # via -r requirements.in
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2021.1
    # via
    #   babel
    #   pandas
pyzmq==22.1.0
    # via
    #   jupyter-client
    #   jupyter-server
    #   notebook
redis==3.5.3
    # via
    #   -r requirements.in
    #   redisbloom
    #   rmtest
redisbloom==0.4.0
    # via -r requirements.in
requests==2.26.0
    # via
    #   jupyterlab-server
    #   requests-unixsocket
requests-unixsocket==0.2.0
    # via jupyter-server
rmtest==0.7.0
    # via redisbloom
send2trash==1.7.1
    # via
    #   jupyter-server
    #   notebook
six==1.16.0
    # via
    #   argon2-cffi
    #   bleach
    #   cycler
    #   grpcio
    #   jsonschema
    #   protobuf
    #   python-dateutil
    #   redisbloom
    #   thrift
sniffio==1.2.0
    # via anyio
soupsieve==2.2.1
    # via beautifulsoup4
terminado==0.10.1
    # via
    #   jupyter-server
    #   notebook
testpath==0.5.0
    # via nbconvert
thrift==0.13.0
    # via fastparquet
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
tqdm==4.61.2
    # via -r requirements.in
traitlets==5.0.5
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-server
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
urllib3==1.26.6
    # via
    #   requests
    #   requests-unixsocket
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
websocket-client==1.1.0
    # via jupyter-server

# The following packages are considered to be unsafe in a requirements file:
# setuptools
